fix issues with whiteout handling
there were two issues in how undocker was handling whiteouts.  When
using the `-o <output directory>` option, that path prefix was not
being applied when processing whiteouts so undocker was for the most
part ignoring the whiteouts.

Undocker was always calling `os.unlink` in order to remove filesystem
paths, which fails when the path being removed is a directory.

Closes #17.